Extract project tile data and devicon data into separate files
The aim of this is to use a ruby script to extract existing data and store
it in separate files. The information is currently in index.html however
updating it is difficult. With the data extracted out, the index.html
file will be dynamically generated using projects.json and icons.txt